Metric preserving data generator (#89)
Implement linearly scaling relationship and attribute generator, in other words distributions of number of attribtues and relationships to insert increases with the size of the graph (ie. the number of entities/vertices grows).
 
This should preserve metrics as the graph is scaled, whereas before the graph became more sparse with size and thus the metrics (normalized degree distribution, etc) would change as well.

Renaming `specificstrategy` to `schemaspecific` to reflect that this code is tied to the configuration that is passed in. In the future, this should use reflection to load a generating class definition along with the configuration, rather than relying on a configuration tag and switching on this in the factory method for data generation methods.

Removes deprecated data generator files, and fixes a bug in the global transitivity metric `compute` query, which is not restricted to compute over `entity` and `relationship` only, whereas attributes were included via relationships before.