Redefine Graph Scale in Data Generator (#97)
Primary goal:
Implements redefinition as graph scale. Previously, in Data Generator, the total number of entities, relationships, and (unique) attribute instances were counted to create the graph size. This was used to decide when to stop generating, and when to profile the graphs that have been generated. 

New definition: want graph scale to be roughly equivalent to "vertices" in a normal graph. To make this compatible with our typed, attributed hypergraph structure with role players (referred to "extended hypergraph"), we use the following new definition of "vertices" and "edges" (let's call it "edge-centric" definition, rather than a "vertex-centric" or "entity-centric" approach): 
Define "edges" as the number of relationship instances.
Define "vertices" as the number of role players + orphan entities + orphan attributes. Intuitively, this captures all _endpoints_ of relationships, plus disconnected entities and attributes. This new definition succinctly quantifies anything acting as an endpoint, be it entity, relationship or attribute.

Caveats: 
* double counts some relationships that are also role players
* promotes attributes and relationships to contribute to graph scale
  * not bad, just different from counting only entities or everything before
Consequences:
* can track quality of data generator by looking at number of orphaned entities and attributes
* track double counting between relationships (edges) and roles (vertices) to see how common it is for relationships to play roles
* Reduces to intuitive definition of "scale" being number of vertices when (hyper)-edges only connected entities together (ie. only entities play roles)