Track roles played in Ignite, data generator performance enhancement (#105)
Because we don't have negation yet, and the prior solution to finding concepts NOT in a relationship yet, resorted to randomly sampling concepts from ignite, then checking in Grakn if these sampled concepts are in the given relationship with the given role.
This was prohibitively slow and scales badly since most entities are not in relationships in the generation phase.

New solution is to add columns to the ignite per-type tables. One column has been added for every relationship type in the schema. When a concept plays a role, we now add it to the big Role Players table (for counting), and append the role player string to the concept's row/relationship column, growing a sort of comma separated list in that cell.

Now when finding concepts of a type not in a specific relationship/role, we query directly in the ingnite table and return the list of concepts not in that role in that relationship already. These sets tend to be small (<100 orphan entities overall in the graph during generation!)