Create attributes identically to Entities and Relationships in Data Generator (#109)
Previously, attributes were treated separately to entities and relationships in the data generator. This means that attributes were created and added to entities/relationships at a fixed rate or a scaling rate.

This PR separates attribute creation into two parts: insert unique attribute values at a fixed rate (same as entities), but these are always created as orphans, just like entities. Then separately, the implicit relationships are created by connecting (attribute instance) <-- @has-attr-value--[@has-attr]--@has-attr-owner--> (concept). This requires a bit more work in the definition of the data generator, but is more flexible and allows us to maintain graph metrics over scale, in the new definition of scale IF we use scaling PDFs for the @has-attr relationships.