Data generator graph tweaks (#123)
Tweaks to the data generators, notably bugs in the role player definitions where the wrong labels (eg. for the role player type instead of the relationship type) were provided.

Also now throws an error in the Data Generator if the provided relationship type for the role player is not a valid relationship type.