Report Generator (#161)
The report generator shares components with the profiler with a slightly different goal.
Report generator:

spawn two separate client & server machines to simulate a realistic client & server setup
execute a set of configuration files, without tracing and only measuring time on the client side. The set of configuration files executed is a subset of profiler, meant to give a more high-level overview of Grakn performance.
A bash script spaws the GCP client and server machines, which each in turn launch the server and report-generator. Each configuration file (representing a scenario) produces a single .json output file, which are aggregated and pulled back onto the originating machine.
formatter is the second component of report which accepts the aggregated json file and produces a formatted text file report.
Minor changes:

adjust scales of data generator
refactor package structures