Track time spent on Generation/Querying/Other and bump deps (#179)
When using data generator in conjunction with profiling or recording, we have a couple of heavyweight operations going on: 
1) querying (ie. profilng how long queries take, the actual goal)
2) generating the graphs to profile agains
3) generating the queries that create the graph (this is the bit using Ignite extensively)

It was a nice to have from #176 to track the proportion of time spent on each. This PR implements this with a new small class in `//common/timer` called `BenchmarkTimer`. It simply sums up time spent on:
1) executing the profiler/recorder queries (everything, including deleting inserted instances) -- corresponds to point 1 above
2) executing `insert` queries that grow the graph (corresponds to point 2 above)
3) everything else, which roughly corresponds to time spent on ignite and generating queries (point 3 above)


--- also, bump dependencies on client-java, graql, and build-tools. These are also going to be auto-updated by `grabl`