handle get requests with vue router history mode (#233)
## What is the goal of this PR?
To allow accessing/sharing direct URLs within the dashboard. Previously accessing a valid link directly would result in 404. Now it works as expected.

## What are the changes implemented in this PR?
within `web-server/src/server.js`, the `connect-history-api-fallback` middleware is used on the express `app` to rewrite the request url of incoming Get requests (which meet certain conditions) to be (the default) `index.html`.