exclude outliers from QueryCard's histogram data (#242)
## What is the goal of this PR?
Previously, the histogram of QueryCard excluded the first repetitions of the query (with the assumption that the first repetition always took significantly longer). Turns out, for many queries that is not the case. As a part of this PR, the _outlier_ spans, if any, regardless of their repetitions number, are excluded. The outlier spans are displayed on the QueryCard, left of the histogram.

## What are the changes implemented in this PR?
`util/math.js` contains the implementation for identifying the outliers within an array of numbers. Moving forward, this file will act as a small math library.

`QueryCard.js` imports from `math.js` and uses `getMedian` and getOutliers` functions for obtaning the `histogramSpans` and `outlierSpans`.