Rename scenarios and update tags (#264)
## What is the goal of this PR?
Dashboard overview's graph titles currently are not aesthetic - ending in `.gql` for instance, or requiring string processing to replace `_` with ` `. 

We can improve this by using the names of benchmark configurations rather than the `graphType`. This PR cleans up the configuration names and the zipkin tags that are stored in ES


## What are the changes implemented in this PR?
Rename config files and graph tags.