Change from IP to internal google name (#283)
## What is the goal of this PR?
Benchmark executions fail when they cannot reach the elasticsearch instance. Using `benchmark.grakn.ai` instead of the explicit external google-managed IP fails for unknown reasons, even when elastic is configured to listen to all network hosts. The workaround for this is to use the internal google VM name (`benchmark-service`) that we do not expect to have change instead of the IP or the URL.
Google internal converts this into an internal URI. This can be also be combined with more strict network host settings in elastic to only allow datacenter-local access to the server.


## What are the changes implemented in this PR?
* Replace direct accesses to the elastic IP with `benchmark-service` for all cases that are not calling the web server endpoints (those are separate HTTPS calls)