Rectify Execution and VM controllers (#288)
## What is the goal of this PR?

The make the code for `ExecutionController` and `VMController` consistent with the implementation of the `SpanController`.

## What are the changes implemented in this PR?
Apart from minor improvements, this PR also contains a bugfix. Within `ExecutionController.updateStatusInternal`, await is no longer used for `downloadLogs`, so that this task takes place in the background allowing `updateStatus`, where `updateStatusInternal` may be invoked from, to respond as soon as the status of the execution has been updated.