Query Generator v1 (#301)
## What is the goal of this PR?
Query generator that consumes a keyspace and generates a set number of queries with varying levels of ambiguity, specificity, number of attributes and role players etc.

The goal is to be able to generate plausible queries according to a schema. Note that a single schema permits many different shapes of data (eg. a marriage between `spouse` roles may in practice only have a maximum of 2 role players, but the schema allows infinite). 

This version chooses how many variables to generate and how many role players and attribtues to add to a query based on geometric series (eg. probability 1, 0.5, 0.25... or 0.5, 0.5*0.3...) etc. This should generate mostly shorter queries but also some longer ones we can choose to keep at a later filtering stage.

The first version should allow generating according to a schema. Implements point one in #300. 

## What are the changes implemented in this PR?
* new package `querygen` that can be used to return a list of `GraqlGet` queries
* subpackages `src` and `test-integration`
* add license headers to some missing files