Query generator v3 - Query Metrics (#305)
## What is the goal of this PR?
Implements part 3 of #300 - a set of methods that convert a query into a numerican vector that can be compared. The next step will be to sample the queries generated to obtain a good spread in this vector space. Hopefully this will ensure that we get a good representation of possible queries for particular schema.

## What are the changes implemented in this PR?
* Implement a stateful `Vectoriser`, which can be used to convert a `QueryBuilder` into a numerical vector
* Chose to operate on `QueryBuilder` rather than the built `GraqlGet` query, as this is easier to implement and doesn't require opening a new transaction to walk the schema for type hierarchy access
* Tests for `Vectoriser` to ensure they are correct